feat(phai): add embeddable read-only sandbox run viewer
Extract the sandbox thread renderer into a shared SandboxThreadView and add a SandboxRunViewer that replays any task run's agent thread read-only, reusing sandboxStreamLogic. Read-only instances are keyed apart from live streams and can never open SSE.
